demo map: cavern room + start-available ore quest (demo map/gather)
The demo map grows to three rooms: camp <-east/west-> area <-north-mouth/
south-mouth-> cavern. The cavern (256x112, authored mask + 4-shade art) is a
safe mine with six nodes (ore x3, bug x2, mushroom) and no beast; the area
gains a second door on its north edge (176,0,16,8) plus the from_cavern spawn,
and the cavern's south mouth (112,104,16,8) returns to the area at
SPAWN_AREA_FROM_CAVERN. Mask rects are matched to the JSON prop list in
canonical order, so the cavern prop list is ordered to match the mask.

gather_ore is now unlock 0 with its desc pointing at the cavern, so the demo
can show gather -> turn-in immediately; crush_heavy still chains off it. The
gather-depletion mask widened u16 -> u32 (19 props total; items.hpp asserts
PROPS_COUNT <= 32).

Tests: zone_test loads the cavern record (extents, safe room, 6 props, spawn)
and round-trips area <-> cavern both ways; quests_test pins the new unlock.
Docs: map room graph + gather table + quest chain updated.

gen-check, host 6916/0, tools 394 OK, 19/19 device suites. Shipping
29490/29696 (206 free), RAM 1916.

+++ src/core/game.hpp
@@ -982,7 +982,7 @@ struct Game {
     // never touches it, so a node stays depleted across a room round-trip).
     // Appended last so a default Game keeps every existing field offset.
     uint8_t items[ITEM_COUNT];
-    uint16_t gatherMask;
+    uint32_t gatherMask;
     // Carve (bead monhun-ardu-prg.3): carvesDone is the successful-carve count
     // this hunt (cap CARVE_MAX; newGame resets it). carveHold mirrors a live
     // PS_CARVE so the over-screen app layer knows the A press belongs to the
